requirements.txt
requests
random
socket
dnspython
threading
rich